fix(review): move regression spec to shared/utils/, follow spec conventions
Convention reviewer: the spec queried by CSS element selector instead
of data-testid, skipped configureTestingModule/compileComponents,
and lived at a loose top-level file in shared/ (not a documented
subdirectory). Moved to shared/utils/, added a data-testid target,
and wrapped creation through the documented TestBed setup. Also adds
a case covering the update path (a signal-driven [class] change),
per general-reviewer feedback that initial-render-only coverage
missed the path production templates actually exercise.

GH-1784
